4차시 과제 답

+++ Assets/Scripts/Enemy.cs
@@ -5,19 +5,17 @@ using UnityEngine;
 
 public class Enemy : MonoBehaviour
 {
-    public int speed;
+    public float speed;
     public Vector2 minAlivePos;
     public Vector2 maxAlivePos;
     private void Update()
     {
         transform.position += transform.rotation * Vector2.up * Time.deltaTime * speed;
 
-        if (transform.position.x < minAlivePos.x || transform.position.y < minAlivePos.y||
+        if (transform.position.x < minAlivePos.x || transform.position.y < minAlivePos.y ||
             transform.position.x > maxAlivePos.x || transform.position.y > maxAlivePos.y)
         {
             Destroy(gameObject);
         }
-
-        
     }
 }
